Preserve specifier characters for printf
Some of our translations use specifier characters like this:

The old timezone: %s

The percentage sign used in the specifier is removed
when we don't restore the IFS to its state of before the
dialog. Therefor, the string we want to print is outputted
like this:

The old timezone: s

The second string specifier character can't be used
in printf functions and therefor the result is a wrongly
formatted output.